roachprod: move agents to agents package
Previously, agent setup and configs were added to the root pkg in roachprod.
This changes adds a bit more organization and moves these packages to an
`agents` package.

Epic: None
Release note: None
