+++ pkg/BUILD.bazel
@@ -322,13 +322,13 @@ ALL_TESTS = [
     "//pkg/revert:revert_test",
     "//pkg/roachpb:roachpb_disallowed_imports_test",
     "//pkg/roachpb:roachpb_test",
+    "//pkg/roachprod/agents/opentelemetry:opentelemetry_test",
     "//pkg/roachprod/blobfixture:blobfixture_test",
     "//pkg/roachprod/cloud:cloud_test",
     "//pkg/roachprod/config:config_test",
     "//pkg/roachprod/failureinjection/failures:failures_test",
     "//pkg/roachprod/install:install_test",
     "//pkg/roachprod/logger:logger_test",
-    "//pkg/roachprod/opentelemetry:opentelemetry_test",
     "//pkg/roachprod/prometheus:prometheus_test",
     "//pkg/roachprod/promhelperclient:promhelperclient_test",
     "//pkg/roachprod/roachprodutil/codec:codec_test",
@@ -1706,6 +1706,9 @@ GO_TARGETS = [
     "//pkg/revert:revert_test",
     "//pkg/roachpb:roachpb",
     "//pkg/roachpb:roachpb_test",
+    "//pkg/roachprod/agents/fluentbit:fluentbit",
+    "//pkg/roachprod/agents/opentelemetry:opentelemetry",
+    "//pkg/roachprod/agents/opentelemetry:opentelemetry_test",
     "//pkg/roachprod/blobfixture:blobfixture",
     "//pkg/roachprod/blobfixture:blobfixture_test",
     "//pkg/roachprod/cloud:cloud",
@@ -1715,14 +1718,11 @@ GO_TARGETS = [
     "//pkg/roachprod/errors:errors",
     "//pkg/roachprod/failureinjection/failures:failures",
     "//pkg/roachprod/failureinjection/failures:failures_test",
-    "//pkg/roachprod/fluentbit:fluentbit",
     "//pkg/roachprod/install:install",
     "//pkg/roachprod/install:install_test",
     "//pkg/roachprod/lock:lock",
     "//pkg/roachprod/logger:logger",
     "//pkg/roachprod/logger:logger_test",
-    "//pkg/roachprod/opentelemetry:opentelemetry",
-    "//pkg/roachprod/opentelemetry:opentelemetry_test",
     "//pkg/roachprod/prometheus:prometheus",
     "//pkg/roachprod/prometheus:prometheus_test",
     "//pkg/roachprod/promhelperclient:promhelperclient",
