+++ pkg/cmd/roachprod/cli/BUILD.bazel
@@ -16,12 +16,12 @@ go_library(
     deps = [
         "//pkg/cmd/roachprod/grafana",
         "//pkg/roachprod",
+        "//pkg/roachprod/agents/fluentbit",
+        "//pkg/roachprod/agents/opentelemetry",
         "//pkg/roachprod/cloud",
         "//pkg/roachprod/config",
         "//pkg/roachprod/errors",
-        "//pkg/roachprod/fluentbit",
         "//pkg/roachprod/install",
-        "//pkg/roachprod/opentelemetry",
         "//pkg/roachprod/roachprodutil",
         "//pkg/roachprod/ssh",
         "//pkg/roachprod/ui",
