+++ pkg/roachprod/BUILD.bazel
@@ -16,13 +16,13 @@ go_library(
         "//pkg/build",
         "//pkg/cli/exit",
         "//pkg/cmd/roachprod/grafana",
+        "//pkg/roachprod/agents/fluentbit",
+        "//pkg/roachprod/agents/opentelemetry",
         "//pkg/roachprod/cloud",
         "//pkg/roachprod/config",
-        "//pkg/roachprod/fluentbit",
         "//pkg/roachprod/install",
         "//pkg/roachprod/lock",
         "//pkg/roachprod/logger",
-        "//pkg/roachprod/opentelemetry",
         "//pkg/roachprod/prometheus",
         "//pkg/roachprod/promhelperclient",
         "//pkg/roachprod/vm",
