+++ pkg/roachprod/agents/fluentbit/files/fluent-bit.yaml.tmpl
