+++ pkg/roachprod/agents/opentelemetry/files/opentelemetry-config.yaml.tmpl
